Copy nodeSelector, tolerations into Jobs
We already copy `nodeSelector` and `tolerations` from hive-operator into
the various controllers. But we weren't copying them into the Jobs for
imageset, provision, and uninstall. This could result in their Pods e.g.
being scheduled to Nodes that can't run them; or being totally
unschedulable.

With this commit, we:
- Add the namespace and name of the running pod as env vars on
hive-controllers.
- Use those env vars to retrieve the hive-controllers Pod spec.
- Copy the nodeSelector and tolerations from that Pod spec down into the
(PodSpecs of the) Jobs for imageset, provision, and uninstall.

HIVE-2496