MachinePools/OpenStack: Fix nil pointer exception
This is *not* a cherry-pick of #2253 / (c6b37ee6 & f59f327d). That
solution revendored installer to pick up the fix from upstream. In older
branches, this would have dragged in too many dependencies, so we
instead fix it "locally" with an explicit nil check.

Note also that the original fix added unit tests. We can't do that here
either because the new tests rely on the OpenStack UT suite being
un-broken [1], which again relied on upstream changes [2][3] we can't pull into
older branches.

[1] https://github.com/openshift/hive/pull/2251
[2] https://github.com/openshift/installer/pull/8187
[3] https://github.com/openshift/installer/pull/8209

HIVE-2476