Skip Konflux builds when possible
This copies the golang (Re2)regular expression S that we use in prow[0]
and leverages the fact that according to [1] on-cell-expression has a
special "files.all" value that contains a list of all the
added/modified/removed/renamed files. It also says that you can use
CEL[2] expressions on those. Thus, we should only run the build if all
the files in "files.all" do not match the aforementioned regular
expression S.

[0] https://github.com/openshift/release/blob/ee0afebfe1ca88b8c35426120cb175fc8d989ffc/ci-operator/jobs/openshift/hive/openshift-hive-master-presubmits.yaml#L83
[1] https://pipelinesascode.com/docs/guide/authoringprs/
[2] https://github.com/google/cel-spec/blob/master/doc/langdef.md#macros

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>