Make MachinePool poll interval tunable
In addition to `Watch()`ing the local MachinePool and ClusterDeployment
objects, the machinepool controller also sets up a poll. Previously the
interval was hardcoded to 30m. With this commit, we retain that as the
default, but allow it to be tuned via a new field:
`HiveConfig.Spec.MachinePoolPollInterval`.

This is a Duration string -- e.g. `"30m"`. If set to zero or less (e.g.
`"0m"`) we won't poll at all.

HIVE-2536