HIVE-2551: AWS deprov job compliant with cred proc
In #2322, in order to fix AWS provisioining in STS mode we pointed the
credential_process configuration to /output/hiveutil. The problem is
that the same configuration is used for both provisioning and
deprovisioning. Thus, when trying to deprovision in STS mode, when it
came time for delegating operation to hiveutil, we'd get an:

    /output/hiveutil not found

This commit fixes it by replicating the /output/hiveutil existance for
deprovisioning in AWS (though it will only get used in STS mode).

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>