Vendor installer and dependencies
In the process of revendoring the installer, a dependency conflict was
found in which OCP updated its direct dependencies to a new backwards
incompatible Prometheus version (0.48.0). The corresponding k8s dependency update
to the new Prometheus version was only present in k8s 0.31.x+. The new
installer version was dependent on the new Prometheus version as well as
the old version of k8s (0.29.x). This dependency on the old version of
k8s was pulled in by hive through the direct dependency on
openshift/generic-admission-server, which depends on
k8s.io/component-base.

To rectify this issue, update k8s.io/component-base to 0.31.x. As a
byproduct, update controller-runtime to be compatible with
component-base 0.31.x.

Resulting changes:
- github.com/openshift/installer v0.9.0-master.0.20240405164943-304af6735c65 => v0.9.0-master.0.20240613201043-5e36f8fb1dde
- k8s.io/code-generator v0.29.2 => v0.30.2
- k8s.io/component-base v0.29.2 => v0.31.0-alpha.2
sigs.k8s.io/controller-runtime v0.17.2 => v0.18.0
- replace k8s.io/client-go => k8s.io/client-go v0.30.1
- replace k8s.io/apimachinery => k8s.io/apimachinery v0.30.2
- github.com/openshift/library-go v0.0.0-20240207105404-126b47137408 => v0.0.0-20240715191351-e0aa70d55678
- github.com/openshift/machine-api-operator
  v0.2.1-0.20230929171041-2cc7fcf262f3 => v0.2.1-0.20240708094115-80301b13a9d2

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>