Forbid kubeconfig exec
Users with write access to the admin kubeconfig Secret for a given
ClusterDeployment should not be able to execute arbitrary code in the
privileged environment in which we run the controllers that use those
Secrets. Funnel all code paths that load such Secrets through a
validator to ensure that the AuthInfos[].Exec path is not used.

HIVE-2485

(cherry picked from commit df1ea18fbe5e57fca5f98383da0feb606b9afdc1)