hiveutil awsprivatelink to be compatbile with CAPI based installer
In the terraform based installer, the worker security group was named
INFRA_ID-worker-sg. However, when the installer transitioned to the CAPI
based installer, the worker security group name became INFRA_ID-node. As
a result, the hiveutil awsprivatelink endpoint command failed to find
the security group.

This change adds the new name in addition to the old one so that both
formats will be found when searching for the worker security group.