MachinePool failure domain logging
We are having lots of problems with MachinePool scaling since we
introduced failure domain matching as the way to tell whether a
generated MachineSet needs to be created, updated, or deleted on the
spoke. This commit adds logging that prints the contents of the
generated and remote failure domains when a mismatch is detected. It
should help with FFDC in the field.