fieldManager: "hive-$controllername"
When debugging CR updates, e.g. via `--show-managed-fields`, hive used
to annoyingly show `manager: manager`, which is not helpful.

With this commit, we've tried to make sure all updates are driven by a
fieldManager with a `hive` prefix:
- For many paths, the prefix will be `hiveN` where `N` is an integer
  unique to the code path that's configuring the fieldManager. We
  did this to make it easier to trace these paths in the code (which can
  otherwise be Very Hard™).
- `hiveutil` commands will show up as either `hiveutil-$subcommand` or
  `hiveN-util-$subcommand`, depending which path they go through to
  create their client.

If you find cases that still show up as `manager` -- or some other
unhelpful string -- please open a bug!

HIVE-1744