Remove existing Secret files in (deprovision) Pod
For Pods with `restartPolicy: OnFailure`, a failed container may be
rerun in the same Pod, which will reuse the same file system as the
initial run. When we project Secrets (for credentials, certs, etc) to
directories in such containers, those writes can fail the second time
around because the file already exists. Fix by removing the file, if it
exists, before we write it.

Note that at the time of this commit, this only affects deprovision
pods:
- imageset pods don't use ProjectToDir
- provision pods have `restartPolicy: Never`

HIVE-2604