MachinePool GCP: fix zone filter
When `MachinePool.Spec.Platform.GCP.Zones` does not explicitly enumerate
the zones to use, we attempt to default to all zones in the configured
region. However, this query was using an overly-broad regex which
matched any zone whose region was a *substring* of the configured
region. This broke when GCP added regions `europe-west10` and
`europe-west12`, which our filter would retrieve when the cluster was in
region `europe-west1`.

The zone list response includes the region as a full URI (e.g.
`https://www.googleapis.com/compute/v1/projects/openshift-hive-dev/regions/europe-west1`).
Fix our regex to match our full region string, anchored by the last `/`
and the end of the string.

HIVE-2610