ClusterVersion: tunable poll interval
Add HiveConfig.Spec.ClusterVersionPollInterval, a time.Duration value
which (plus random jitter of 0-1s) will be used to requeue
ClusterDeployments for the clusterversion controller, which is
responsible for setting `hive.openshift.io/version*` labels on
ClusterDeployment objects based on the ClusterVersion object on the
spoke.

If unset or zero, this is disabled, resulting in the legacy behavior: we
will reconcile on controller startup, ClusterDeployment changes, or (I
think) the default controller-runtime interval of 1000s.

HIVE-2619