GCP PrivateLink to use pre-existing psc subnet.
Prior to this change, GCP PrivateLink created a subnet in the cluster
VPC to be used by the service attachment. To do so, it queried the VPC
network from the api and used the returned selfLink for creating the
subnet. However, when using BYO VPCs, the network name is not guaranteed
to match the name created by the installer, so the privatelink
controller would fail to find it. Furthermore, in a Shared VPC
configuration, Hive is unable to create the subnet in the host project.

This change adds new fields to the clusterDeployment spec for GCP
PrivateLink configuration to enable a pre-existing subnet to be
configured. In order to accommodate Shared VPC, it also enables the
configuration of the project name. As a result, GCP PrivateLink can now
deploy using BYO VPC in addition to Shared VPC configurations.