Revendor installer to 0.91.0
This is a special revendor.

A prior commit (#2460 / c7581516) included installer as a dependency in
our apis/ package for the first time. This was causing problems for
downstream consumers seemingly because the pseudo-version generated by
`go mod` represented a years-old tag with a recent commit, resulting in
excessive churn in the go sum database -- see the referenced bug report.

This commit revendors to the (current) latest `master` where a tag has
been created to make this pseudo-version less gappy, hopefully
eliminating this problem for downstreams.

OCPBUGS-42448

HIVE-2512