No-op revendor installer
This is a special revendor.

A prior commit (#2463 / d8ebe585) included installer as a dependency in
our apis/ package for the first time. This was causing problems for
downstream consumers seemingly because the pseudo-version generated by
`go mod` represented a years-old tag with a recent commit, resulting in
excessive churn in the go sum database -- see the referenced bug report.

Now a tag has been created at that commit to make this pseudo-version
less gappy, hopefully eliminating this problem for downstreams.

However, when I tried to revendor to that tag, go replaced it with a
pseudo-version... and then complained that we shouldn't use a
pseudo-version when there's a tag available!! (Shakes fist.)

So this commit revendors to the _next_ commit, which should mean that a
pseudo-version is the only way to reference it. The delta on that commit
is null for vendoring purposes (the changed file is not one that makes
its way into the vendor/ directory) so this is just a bump of
pseudo-versions.

OCPBUGS-42448

HIVE-2512