modcheck -f phase 1: fix `require`s
Our modcheck tool looks at the go.mod files in the repository root and
the apis/ directory, and fails if any entries in the latter exist in the
former at a different version. Heretofore, such failures would have to
be addressed by manually updating the apis/go.mod to sync those
versions.

With this change, we add a `-f` flag which asks modcheck to `f`ix
discrepancies it encounters.

In this first phase, we add the ability for modcheck to fix `require`
entries, as these are by far the most common discrepancies we encounter.

*Further work is required to make the tool fix `exclude` and `replace`
entries.*

The tool will succeed (exit zero) if the files are in sync when it
completes -- whether because they started that way, or because it fixed
everything it encountered.

The tool will fail (exit nonzero) if it encounters a processing error;
or if the files are out of sync when it completes -- even if *partially*
fixed.

Put another way: barring errors, the exit code is the same as if
modcheck were run again without `-f`.

HIVE-2624