Upgrade deps to k8s 1.31
NOTE: At this level, k8s.io/apis started using an empty `groupName`
annotation, which broke the version of controller-gen we were using to
generate CRDs. This prompted an upgrade to controller-tools v0.16.1.

NOTE: This also prompted an upgrade to controller-runtime v0.19.0, which
broke some things, which will be fixed in a subsequent commit.

HIVE-2616