controller-gen: use vendor/, not build-machinery-go
The build-machinery-go project is suffering from bit rot, which is
making it difficult to upgrade the version of controller-gen we use to
generate CRDs, which we're going to need to do in order to upgrade k8s
dependencies (see HIVE-2616).

With this change, we stop using the piece of build-machinery-go that
deals with controller-gen, and instead install it directly from our
vendor directory.

HIVE-2626