MachinePool: Generated MachineLabels are not "owned"
The machinepool controller tracks labels, machineLabels, and taints in
the MachinePool.Status (OwnedLabels, OwnedMachineLabels, OwnedTaints) so
that, when the user removes them from the spec, we can idempotently know
which ones to remove from the spoke MachineSet. This is to avoid
removing any of those things if they were added directly by the user on
the spoke rather than through the MachinePool.

Note that, in the case of a conflict (the same key in both the
MachinePool and added directly on the spoke), the MachinePool would
"win".

However, when MachineLabels were implemented (#2420 / 315aa266) we
needed special handling to avoid stomping on *generated* labels -- those
produced by the installer code that generates the MachineSet -- because
these have special meaning and function to the controllers that manage
the MachineSets. Thus, user-specified machineLabels conflicting with
generated ones would be ignored -- the resulting MachineSet would retain
the generated values.

But there was a bug: we were still adding the keys to
OwnedMachineLabels; so if they were subsequently *removed* from
MachinePool.Spec.MachineLabels, we would attempt to *remove* them from
the MachineSets. This resulted in webhook errors creating the
MachineSets, which is good, because otherwise the MachineSets would have
been broken in some more subtle way that would be difficult to detect.

With this fix, we start tracking which machineLabels are generated, and
avoid adding them to OwnedMachineLabels. So, as originally intended,
MachinePool.Spec.MachineLabels whose keys conflict with generated labels
are truly ignored.

HIVE-2320