Hibernation: Requeue without error when ClusterSync doesn't exist yet
When installation finishes, the clustersync and hibernation controllers
race to start their business. But the latter needs to make use of the
ClusterSync object. When the clustersync controller hasn't created it
yet, the hibernation controller needs to requeue and wait for it.

Previously we were just letting the 404 from Get(ClusterSync) to bubble
up; but this counts against the reconciler error budget for which alerts
may be triggered. Since this is an expected scenario, we change this
code path to do a non-error immediate requeue instead.

HIVE-2611