Nix yaml-patch
HIVE-2397:

Remove uses of `github.com/krishicks/yaml-patch`, which is unmaintained.

Replace by wrapping `gopkg.in/evanphx/json-patch.v4` with a new function to:
- Convert YAML to JSON
- Convert `PatchEntity` -- see below -- to a patch suitable to...
- ...use json-patch to patch the JSON
- Convert back to YAML

This allowed us to convert most of the code previously using yaml-patch
with relatively little disruption to the code flow.

In anticipation of future work, we decided to reuse `PatchEntity`. This
is a hive API type originally designed to patch install-config templates
for ClusterPool clusters; but it is suitable for this purpose since it
represents [RFC6902](https://datatracker.ietf.org/doc/html/rfc6902)
patches. However...

HIVE-2214:

In this effort, it became necessary to add support for complex
(non-string) types to `PatchEntity`. We did this by adding a new member,
`ValueJSON`, which accepts a string representing encoded JSON. `Value`
can still be used for string types.

Also:
In passing, we noticed that PatchEntity.Op, which is restricted to a
discrete set of values per the RFC, was being validated by webhook.
We've changed this to use enum schema validation instead.

HIVE-2397
HIVE-2214