Add ClusterDeploymentLabelSelector support
This commit adds the labelSelector customization for some metrics. This
feature can be opted into via HiveConfig, and the metrics, for which this
customization is configured, will only be reported if the label query
matches the labels on the related clusterdeployment.

To get around cyclic import cycle, the definition and registration of
all supported metrics from clusterDeployment, clusterProvision and
clusterDeprovision controller have been moved to metrics controller.