Azure: Pass BaseDomainResourceGroupName to destroyer
If an Azure spoke was provisioned with a custom
BaseDomainResourceGroupName, DNS records were being leaked when
destroying because we weren't passing that information through to the
deprovisioner.

This change adds a new API field,
ClusterDeployment.Spec.ClusterMetadata.Platform.Azure.BaseDomainResourceGroupName
(and a corresponding passthrough field in the ClusterDeprovision API).
When set, we will pass this value through to the deprovisioner.

We will glean this value from the metadata produced by the installer and
save it into the ClusterDeployment. However, in certain scenarios
(adoption, upgrade to/past this commit with pre-existing CDs) it may not
be set, and we will not attempt to discover it. In these cases, manual
cleanup may still be necessary.

HIVE-2651