ClusterSync: Warn on resource conflicts
Add a new ClusterSync.Status field, ResourceConflicts, and a
corresponding printColumn visible when `-o wide` is used.

When a single resource (GVK + name [+ namespace if ns-scoped]) is
mentioned more than once in the `resources` sections (not `patches` or
`secretMappings`) of the [Selector]SyncSets matching a
ClusterDeployment, we will add a human-readable message to this status
field indicating the resource and the [Selector]SyncSets in which it is
mentioned.

Example message:

```
"Resource {v1 Namespace openshift-logging } is mentioned by 2 [Selector]SyncSets: mygroup, osd-logging1"
```

HIVE-2634