syncset: support templates in `patches`
HIVE-1748 / #2192 added the ability to flip a switch in a
[Selector]SyncSet that causes hive to process golang
[text/template](https://pkg.go.dev/text/template)s in leaf values of
objects in the [S]SS's `resources` section.

This does a similar thing for JSON patch strings in the `patches`
section. For example:

```yaml
apiVersion: hive.openshift.io/v1
kind: SyncSet
metadata:
  name: paramss
spec:
  ...
  enablePatchTemplates: true
  ...

  patches:
  - apiVersion: v1
    kind: ConfigMap
    name: foo
    namespace: default
    patch: |-
      {"metadata": {"annotations": {"test": "{{ fromCDLabel "hive.openshift.io/cluster-platform" }}" }}}
    patchType: merge
```

Note that the data type of `patches[].patch` is `string`, so internal
quotes must be escaped somehow (in the above example, it's via `|-`).
But the value of the patch _once the template is rendered_ must be valid
JSON, e.g. with double quotes surrounding string values. This can make
the quoting in the unrendered patch look syntactically odd.

HIVE-2669