GCP: Hibernation: Support SSDs
It would seem as though, when stopping an instance with SSDs, one is
required to specify whether their contents should be preserved or
discarded, despite what the documentation [1] says about defaulting that
behavior to the latter. This will result in a failure to stop instances,
with the following message appearing in the ClusterDeployment's Hibernating condition:

```
Error 400: VM has a Local SSD attached but
an undefined value for `discard-local-ssd`. If using gcloud, please add `--discard-local-ssd=false`
or `--discard-local-ssd=true` to your command.
```

This change adds
ClusterDeployment.Spec.Platform.GCP.DiscardLocalSsdOnHibernate, a
_pointer_ to a bool that we will pass through to the GCP API if
provided. The default behavior is as before: when the field is not
populated, we will not pass the option to the GCP API at all, which
continues to be fine for non-SSD-having VMs.

For usability purposes, we allow this field to be edited (most spec
fields are immutable, enforced by validating webhook).

[1] https://cloud.google.com/compute/docs/disks/local-ssd#stop_instance

HIVE-2693