bundle-gen.py: build and push public images
Our OperatorHub bundle generator validates the discovered/specified
operator image by querying the repository to see if it exists.
Previously if it did not exist, we would simply fail. With this change,
we instead attempt to build and push it.

This also changes the default repository from AppSRE's to hive's.

...and addresses a latent syntax error in a regex used to generate
semvers from MCE branch names.

HIVE-2766