HIVE-2788: MachinePool/GCP: Remove DiskType enum
GCP seems to have added some disk types since we wrote our API for
MachinePool.Spec.Platform.GCP.OSDisk.DiskType with an enum restricting
it to two values (`pd-standard` and `pd-ssd`).

The proximate need is to be able to stand up C3 instances such as
`c3-standard-192-metal`, which require a disk type of
`hyperdisk-balanced`.

However, for future compatibility -- i.e. against the advent of GCP
adding more disk types down the road -- this change simply removes the
restriction, allowing the user to enter any string value. If they get it
wrong, it just means it will be GCP rather than k8s that bounces the
request.

Note that absorbing this change should only be a matter of reinstalling
CRDs (specifically MachinePool). Controller code is not affected, here
or downstream.