HIVE-2870: hiveadmission: proper authn to the KAS
This change corrects the hiveadmission service's authentication to the
hub's kube-apiserver by injecting the correct CA bundle into the
ValidatingWebhookConfigurations (VWHCs) when running on OpenShift.

Previously we were relying on the
`service.beta.openshift.io/inject-cabundle` annotation. It turns out
this injects the *wrong* bundle for VWHCs (though, accidentally, the
right one for the aggregated APIService). Our webhooks still worked
because the KAS has special handling, ignoring the VWHC's caBundle
entirely, for clients using the aggregated API service, which
hiveadmission does. However, OpenShift's kube-apiserver-operator (KASO)
has a webhook validator that lacks this special handling (see
OCPBUGS-50978) and thus generates error logs. While these errors are
harmless and do not reflect any functional problem, they can be
annoying/distracting for customers.

Modern (all currently-supported) versions of OpenShift create ConfigMaps
in every namespace containing the service CA cert, which is needed by the
APIService to authn to the KAS, and the kube root CA cert, which is
needed in the VWHCs by KASO (though not by the KAS) to validate the
webhooks. So this change:
- Removes the `inject-cabundle` annotation from the APIService and all
  VWHCs.
- Adds a code path for OpenShift clusters to look up these ConfigMaps
  and inject the respective certs into the respective objects
  explicitly.

The logic for non-OpenShift clusters remains the same.