HIVE-2302: Pass metadata.json through opaquely
Well, mostly.

Previously any time installer added a field to metadata.json, we would
need to evaluate and possibly add a bespoke field and code path for it
to make sure it was supplied to the destroyer at deprovision time.

With this change, we're instead offloading it verbatim to a new Secret
in the ClusterDeployment's namespace, referenced from a new field:
ClusterDeployment.Spec.ClusterMetadata.MetadataJSONSecretRef.

Instead of building the installer's ClusterMetadata structure for the
destroyer with individual fields from the CD's ClusterMetadata, we're
unmarshaling it directly from the contents of that Secret.

(Except in some cases we have to scrub/replace credentials fields -- see
HIVE-2804 / #2612)

For legacy clusters -- those created before this change -- we attempt to
retrofit the new Secret based on the legacy fields. This is best effort
and may not always work. If this results in a hanging deprovision due to
a missing field, the workaround is to modify the contents of the Secret
to add it in; then kill the deprovision pod and the next attempt should
pick up the changes. (If the result is a "successful" deprovision with
leaked resources, the only workaround is to clean up the infra manually.
Sorry.)